+++ src/components/admin/adminList/index.vue
@@ -3,22 +3,20 @@ import { NButton, NGrid, NGi, NIcon } from 'naive-ui'
 import { VxeButtonEvents, VXETable, VxeTableEvents, VxeTableInstance } from 'vxe-table'
 import { TrashOutline, Add } from '@vicons/ionicons5'
 import { EditRegular } from '@vicons/fa'
-import { onMounted, reactive, ref } from 'vue'
 import {
-  addAreaApi,
-  editAreaByNameApi,
-  editCategoryByNameApi,
-  listAreaByPageApi,
-  removeAreaBatchByIdsApi,
-  removeAreaByIdApi
-} from '@service/property'
+  addAdminApi,
+  editAdminApi,
+  listAdminByPageApi,
+  removeAdminBatchByIdsApi,
+  removeAdminByIdApi
+} from '@service/adminList'
+import { onMounted, reactive, ref } from 'vue'
 import underLine from '@/assets/underLine.png'
-import AreaQueryParams = API.AreaQueryParams
 
 const underLinePng = underLine
 
 // 页面数据
-const propertyData = reactive({
+const adminData = reactive({
   // 表格数据
   tableData: [],
   tablePage: {
@@ -28,48 +26,54 @@ const propertyData = reactive({
   }
 })
 
-// 查询地区数据的条件参数
-const params = reactive<API.AreaQueryParams & API.PageParams>({
-  pageSize: propertyData.tablePage.pageSize,
-  page: propertyData.tablePage.currentPage,
-  areaName: ''
+// 查询管理员数据的条件参数
+const params = reactive<API.AdminQueryParams & API.PageParams>({
+  pageSize: adminData.tablePage.pageSize,
+  page: adminData.tablePage.currentPage,
+  username: ''
 })
 
-// 修改地区数据的参数
-const updateParams = reactive<API.AreaParams>({
-  areaId: 0,
-  areaCode: '',
-  areaDescription: '',
-  areaName: ''
+// 新增管理员参数
+const addParams = reactive<API.AdminParams>({
+  username: '',
+  passWord: '',
+  phone: ''
+})
+
+// 修改管理员参数
+const updateParams = reactive<API.AdminParams>({
+  userId: 0,
+  username: '',
+  passWord: '',
+  phone: ''
 })
 
 // 根据参数分页查询所有数据
-const findAreaSelectPage = async () => {
-  const res = await listAreaByPageApi(params)
-  propertyData.tableData = res.data! as any
-  propertyData.tablePage.total = res.count! as any
-  console.log(propertyData.tableData)
+const findAdminSelectPage = async () => {
+  const res = await listAdminByPageApi(params)
+  adminData.tableData = res.data! as any
+  adminData.tablePage.total = res.count! as any
+  console.log(adminData.tableData)
 }
 
-// 搜索地区名
-function changeAreaName(event: any) {
-  params.areaName = event.value
+// 搜索管理员名
+function changeAdminName(event: any) {
+  params.username = event.value
   console.log(event.value)
-  findAreaSelectPage()
+  findAdminSelectPage()
 }
 
 // 重置参数
 const clickEvent = () => {
-  params.areaName = ''
-  // propertyData.option.area = ''
-  findAreaSelectPage()
+  params.username = ''
+  findAdminSelectPage()
 }
 
 // 页码切换
 function handleChangePage() {
-  params.pageSize = propertyData.tablePage.pageSize
-  params.page = propertyData.tablePage.currentPage
-  findAreaSelectPage()
+  params.pageSize = adminData.tablePage.pageSize
+  params.page = adminData.tablePage.currentPage
+  findAdminSelectPage()
 }
 
 // 添加框是否显示
@@ -77,47 +81,50 @@ const isShow = ref<boolean>(false)
 
 // 修改框是否显示
 const isShow2 = ref<boolean>(false)
-// 添加地区名
-const inputAreaName = ref<string>('')
 
 // 表格
 const xTable1 = ref<VxeTableInstance>()
 
-// 添加地区数据
+// 添加管理员数据
 const insertEvent = async () => {
-  await addAreaApi(inputAreaName.value)
+  await addAdminApi(addParams)
   isShow.value = false
-  await findAreaSelectPage()
+  await findAdminSelectPage()
 }
 
-// 修改分类数据事件
+// 修改管理员数据事件
 const updateEvent = async () => {
   // 修改参数
-  updateParams.areaName = inputAreaName.value
-  await editAreaByNameApi(updateParams)
+  await editAdminApi(updateParams)
   isShow2.value = false
-  await findAreaSelectPage()
+  await findAdminSelectPage()
 }
 
-function handleAddArea(event: any) {
-  inputAreaName.value = ''
+function handleAddAdmin(event: any) {
+  addParams.username = ''
+  addParams.passWord = ''
+  addParams.phone = ''
   isShow.value = true
 }
 
-function handleEditArea(row: any) {
+function handleEditAdmin(row: any) {
   // 修改参数id
-  updateParams.areaId = row.areaId as number
+  updateParams.userId = row.userId as number
+  updateParams.username = row.userName as string
+  updateParams.passWord = ''
+  updateParams.phone = ''
+
   isShow2.value = true
 }
 
 // 表格中的删除
 const removeEvent = async (row: any) => {
-  const AreaId = row.areaId as number
+  const UserId = row.userId as number
   const type = await VXETable.modal.confirm('您确定要删除该数据?')
   if (type === 'confirm') {
-    const res = await removeAreaByIdApi(AreaId)
+    const res = await removeAdminByIdApi(UserId)
     if (res.success) {
-      await findAreaSelectPage()
+      await findAdminSelectPage()
       await VXETable.modal.message('删除成功')
     } else {
       await VXETable.modal.message('删除失败')
@@ -126,27 +133,27 @@ const removeEvent = async (row: any) => {
 }
 
 // 定义批量删除的数组
-const areaArray: number[] = []
+const adminArray: number[] = []
 
 // check全选
 const selectAllChangeEvent: VxeTableEvents.CheckboxAll = () => {
   // 置空数组
-  areaArray.length = 0
+  adminArray.length = 0
   const $table = xTable1.value as VxeTableInstance
   const records = $table.getCheckboxRecords()
   records.forEach((item) => {
-    areaArray.push(item.areaId)
+    adminArray.push(item.userId)
   })
 }
 
 // check单选
 const selectChangeEvent: VxeTableEvents.CheckboxChange = () => {
   // 置空数组
-  areaArray.length = 0
+  adminArray.length = 0
   const $table = xTable1.value as VxeTableInstance
   const records = $table.getCheckboxRecords()
   records.forEach((item) => {
-    areaArray.push(item.areaId)
+    adminArray.push(item.userId)
   })
 }
 
@@ -154,9 +161,9 @@ const selectChangeEvent: VxeTableEvents.CheckboxChange = () => {
 const deleteEvent = async () => {
   const type = await VXETable.modal.confirm('您确定要删除该数据?')
   if (type === 'confirm') {
-    const res = await removeAreaBatchByIdsApi(areaArray)
+    const res = await removeAdminBatchByIdsApi(adminArray)
     if (res.data) {
-      await findAreaSelectPage()
+      await findAdminSelectPage()
       await VXETable.modal.message('删除成功')
     } else {
       await VXETable.modal.message('删除失败')
@@ -165,31 +172,31 @@ const deleteEvent = async () => {
 }
 
 onMounted(() => {
-  findAreaSelectPage()
+  findAdminSelectPage()
 })
 </script>
 
 <template>
   <div class="bg_box">
     <div class="bg_box2">
-      <h3>地区管理</h3>
-      <img class="text_png1" :src="underLinePng" alt="下划线" />
+      <h3>管理账号列表</h3>
+      <img class="text_png1" :src="underLinePng" />
       <n-grid :cols="4" x-gap="24">
         <n-gi span="4">
           <vxe-toolbar>
             <template #buttons>
               <vxe-input
                 placeholder="输入关键词搜索"
-                v-model="params.areaName"
+                v-model="params.username"
                 type="search"
-                @blur="changeAreaName"
+                @blur="changeAdminName"
               ></vxe-input>
 
-              <n-button ghost type="info" style="margin-left: 5px" @click="clickEvent">
+              <n-button ghost type="info" style="margin-left: 70px" @click="clickEvent">
                 重置
               </n-button>
 
-              <n-button type="info" style="margin-left: 5px" @click="handleAddArea">
+              <n-button type="info" style="margin-left: 350px" @click="handleAddAdmin">
                 <template #icon>
                   <n-icon>
                     <Add />
@@ -197,7 +204,7 @@ onMounted(() => {
                 </template>
                 添加
               </n-button>
-              <n-button type="error" style="margin-left: 5px" @click="deleteEvent">
+              <n-button type="error" style="margin-left: 20px" @click="deleteEvent">
                 <template #icon>
                   <n-icon>
                     <TrashOutline />
@@ -217,22 +224,23 @@ onMounted(() => {
               ref="xTable1"
               size="small"
               align="center"
-              :data="propertyData.tableData"
+              :data="adminData.tableData"
               :export-config="{}"
               :row-config="{ isHover: true }"
               @checkbox-all="selectAllChangeEvent"
               @checkbox-change="selectChangeEvent"
             >
               <vxe-column type="checkbox" width="60"></vxe-column>
-              <vxe-table-column field="areaId" title="序号" width="140"></vxe-table-column>
-              <vxe-table-column field="areaName" title="地区名称" width="350"></vxe-table-column>
+              <vxe-table-column field="userId" title="序号" width="100"></vxe-table-column>
+              <vxe-table-column field="userName" title="账号名称" width="300"></vxe-table-column>
+              <vxe-table-column field="createdAt" title="添加日期" width="300"></vxe-table-column>
               <vxe-column show-overflow title="操作">
                 <template #default="{ row }">
                   <div>
                     <n-button text>
                       <template #icon>
                         <n-icon :depth="3">
-                          <EditRegular @click="handleEditArea(row)" />
+                          <EditRegular @click="handleEditAdmin(row)" />
                         </n-icon>
                       </template>
                     </n-button>
@@ -255,8 +263,8 @@ onMounted(() => {
       <n-grid :cols="4" x-gap="24">
         <n-gi span="4">
           <vxe-pager
-            v-model:current-page="propertyData.tablePage.currentPage"
-            v-model:page-size="propertyData.tablePage.pageSize"
+            v-model:current-page="adminData.tablePage.currentPage"
+            v-model:page-size="adminData.tablePage.pageSize"
             :layouts="[
               'PrevJump',
               'PrevPage',
@@ -266,28 +274,41 @@ onMounted(() => {
               'FullJump',
               'Total'
             ]"
-            :total="propertyData.tablePage.total"
+            :total="adminData.tablePage.total"
             background
             @page-change="handleChangePage"
           >
           </vxe-pager>
         </n-gi>
       </n-grid>
-      <!--      新增地区弹出框-->
+      <!--      新增管理员弹出框-->
       <vxe-modal
         v-model="isShow"
         destroy-on-close
         min-height="900"
         min-width="900"
         resize
-        title="添加地区"
-        width="340"
-        height="230"
+        title="管理员账号添加"
+        width="380"
+        height="330"
       >
         <template #default>
           <div>
             <div class="input_text">
-              <vxe-input v-model="inputAreaName" placeholder="输入新增地区名称"></vxe-input>
+              <div>
+                <text>账号名字</text>
+                <vxe-input v-model="addParams.username" placeholder="输入账户名称"></vxe-input>
+              </div>
+
+              <div>
+                <text>手机号</text>
+                <vxe-input v-model="addParams.phone" placeholder="输入新的手机号"></vxe-input>
+              </div>
+
+              <div>
+                <text>密码</text>
+                <vxe-input v-model="addParams.passWord" placeholder="输入新的密码"></vxe-input>
+              </div>
             </div>
             <div class="input_button">
               <vxe-button
@@ -314,20 +335,33 @@ onMounted(() => {
         min-height="900"
         min-width="900"
         resize
-        title="编辑地区"
-        width="340"
-        height="230"
+        title="管理账号编辑"
+        width="380"
+        height="330"
       >
         <template #default>
           <div>
             <div class="input_text">
-              <vxe-input v-model="inputAreaName" placeholder="输入修改后的名称"></vxe-input>
+              <div>
+                <tex>手机号</tex>
+                <vxe-input v-model="updateParams.phone" placeholder="输入新的手机号"></vxe-input>
+              </div>
+
+              <div>
+                <text>密码</text>
+                <vxe-input v-model="updateParams.passWord" placeholder="输入新的密码"></vxe-input>
+              </div>
+
+              <div>
+                <text>确认密码</text>
+                <vxe-input v-model="updateParams.passWord" placeholder="请确认新的密码"></vxe-input>
+              </div>
             </div>
             <div class="input_button">
               <vxe-button
                 content="取消"
                 status="info"
-                @click="isShow2 = false"
+                @click="isShow = false"
                 class="button_left"
               ></vxe-button>
               <vxe-button
@@ -347,33 +381,25 @@ onMounted(() => {
 <style lang="scss" scoped>
 .bg_box {
   display: flex;
-  margin-top: 50px;
-  margin-bottom: 50px;
-  margin-left: 20px;
-  @media screen and (min-width: 320px) and (max-width: 480px) {
-    margin: 0;
-  }
 
   .bg_box2 {
-    float: right;
-    height: 625px;
-    width: 1080px;
+    height: 80%;
+    width: 80%;
+    margin: 2%;
+    margin-left: 80px;
     padding-left: 3%;
-    padding-right: 1%;
-    padding-bottom: 4%;
+    padding-right: 3%;
+    padding-top: 1%;
+    padding-bottom: 5%;
     border-radius: 15px;
     box-shadow: 0 4px 8px 0 rgba(0, 0, 0, 0.2), 0 6px 20px 0 rgba(0, 0, 0, 0.19);
-    @media screen and (min-width: 320px) and (max-width: 480px) {
-      width: auto;
-    }
 
     .text_png1 {
       position: absolute;
-      height: 17px;
-      width: 144px;
-      left: 395px;
-      bottom: 575px;
-      display: none;
+      height: 16px;
+      width: 150px;
+      left: 335px;
+      bottom: 585px;
     }
 
     .table1 {
@@ -395,11 +421,12 @@ onMounted(() => {
     .input_button {
       margin-left: 30px;
       .button_left {
-        width: 90px;
+        width: 110px;
         margin-right: 45px;
       }
       .button_right {
-        width: 90px;
+        width: 110px;
+        margin-left: 20px;
       }
     }
   }
